added oauth2

+++ src/main/java/com/stardusted1/Sendicate/app/core/repositories/TransmitterRepository.java
@@ -3,5 +3,5 @@ package com.stardusted1.Sendicate.app.core.repositories;
 import com.stardusted1.Sendicate.app.core.cargo.Transmitter;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransmitterRepository extends CrudRepository<Transmitter, Integer> {
+public interface TransmitterRepository extends CrudRepository<Transmitter, Long> {
 }
